Feature6 dodanie spring security

+++ src/main/java/com/gym/gym/service/LockerService.java
@@ -1,33 +1,28 @@
 package com.gym.gym.service;
 
 import com.gym.gym.dto.LockerDto;
-import com.gym.gym.entity.Locker;
+import com.gym.gym.exception.exceptions.EntityNotFoundException;
+import com.gym.gym.exception.messages.ExceptionMessages;
 import com.gym.gym.mapper.LockerMapper;
 import com.gym.gym.repository.LockerRepository;
 import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@AllArgsConstructor
-@Transactional
 @Service
+@Transactional
+@RequiredArgsConstructor
 public class LockerService {
     private final LockerRepository lockerRepository;
+
 
     public List<LockerDto> getAllLockers() {
         return lockerRepository.findAll()
                 .stream()
-                .map(LockerMapper::toDto)
-                .collect(Collectors.toList());
-    }
-
-    public List<LockerDto> getAllAvailableLockers() {
-        return lockerRepository.findAll()
-                .stream()
-                .filter(Locker::isAvailable)
                 .map(LockerMapper::toDto)
                 .collect(Collectors.toList());
     }
@@ -37,6 +32,10 @@ public class LockerService {
     }
 
     public void delete(Long id) {
+
+        if (lockerRepository.findById(id).isEmpty()){
+            throw new EntityNotFoundException("Locker", id);
+        }
         lockerRepository.deleteById(id);
     }
